grub: correct efi_removable_quirk()
Quirk for incomplete (U)EFI implementations (such as EFI implemented by
U-Boot and some vendor firmware). With these implementations, the firmware
may fail to save boot entries, or that the EFI variable store was made
read-only.

Installing GRUB with the --force-extra-removable or --removable flags will
help workaround unbootable systems, though the former is not an option
when the variable store is read-only.

> What about --no-nvram?

No. GRUB still installs the EFI image as grub$ARCH.efi if --no-nvram is
set, but for systems without writable variable or boot entry stores, the
firmware would fail to discover and boot from a non-removable path.

> The workaround.

* --force-extra-removable makes GRUB install a generic BOOT$ARCH.EFI in
  addition to the standard grub$ARCH.efi (along with an EFI boot entry).
* --removable makes GRUB install *only* a generic BOOT$ARCH.EFI, it also
  prompts GRUB to skip writing to the EFI variable store.