plasma-desktop: drop possible mis-set Flat mouse acceleration profile
Revert upstream commit fb8a655a7376 ("If no config option for accel
profile is selected, use default from Xorg").

With libinput 1.26.2, a "flat" acceleration profile was set by default for
all input devices. This is probably not correct as libinput still
indicates that "adaptive" is the default.

Having a "flat" default makes TrackPoints painfully unusable after login
(whereas in SDDM they work as expected).

The user may still select a "flat" profile via the KCM module, save for
the issue where their X configuration may not match with that of KDE's
on first boot (before this reverted commit). This is really much less of
a problem for most users.