glibc: skip empty postrm generation
dpkg trying to run postrm of the HWCAPS enabled glibc causes a hazard
when downgrading to glibc in Core 11. It will try to load then-deleted
glibc installed in HWCAPS subdirectories after dpkg removed the contents
of glibc in Core 12.

Not installing an empty does work around this. We tried various possible
solutions but none of them works.