gcc: backport a patch to fix libgcc_s.so.1 w/ executable stack on MIPS
On MIPS, a #include clause is wrongly commented, which leads to the
.note.GNU-stack section gets omitted in some .o files built from .S
sources, thus the resulting libgcc_s.so.1 gets its GNU_STACK section set
to executable (which is forbidden by newest systemd+kernel for some
system services).

Backport a patch that fixes this issue by restoring the #include clause.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>