krb5: fix FTBFS because of GCC 14
GCC 14 defaults to -Werror=implicit-function-declaration, which breaks
the detection of krb5_cv_attr_constructor_destructor in configure
script (which uses unlink() w/o including unistd.h).

As these compiler attributes are known to be properly working on GCC +
Linux, override the result to "yes,yes" to prevent the broken detection.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>